Move setting validation from package `manager` to package `datastore`
